# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List roles (scoped to caller's tenant + system roles)
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateRole
      tags: [roles]
      summary: Update role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    delete:
      operationId: deleteRole
      tags: [roles]
      summary: Delete role
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Right:
      type: string
      enum:
        - manage_profiles
        - assign_roles
        - manage_roles
        - manage_org_units
        - create_checklists
        - edit_checklists
        - delete_checklists
        - view_reports
        - manage_checklist_settings
...
    Role:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
          nullable: true
        name:
          type: string
        level:
          type: integer
        isSystem:
          type: boolean
        rights:
          type: array
          items:
            $ref: "#/components/schemas/Right"
      required: [id, name, level, isSystem, rights]
    CreateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        level:
          type: integer
        rights:
          type: array
          items:
            $ref: "#/components/schemas/Right"
      required: [name, level, rights]
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        level:
          type: integer
        rights:
          type: array
          items:
            $ref: "#/components/schemas/Right"
...